Mejoras en la visualizacion del contenido , y lectura del codigo Fuente
Este commint serapa la logica Javascript de la Html en archivos separados para tener un mejor control del codigo y una mejor lectura del mismo .

se añade a la vista de facturas  un tooltip mejorado visualmente, ahora los checkbox estaran deshabilitados si la factura contiene algun error , ademas si hay algun error en la factura  el sistema mostrara toda la linea en ROJO  y en el TOOLTIP mostrara la descripccion de los errores encontrados.
