+++ resources/views/FacturacionElectronica/Facturas/index.blade.php
@@ -12,6 +12,7 @@
 
 @section('content')
 
+
     <div class="col-12"><h3> Por favor seleccione un rango de fechas para comenzar con la busqueda.</h3></div>
     <br>
     {!! Form::open(array('url'=>'fe', 'method'=>'GET', 'autcomplete'=>'off', 'role'=>'search', 'id' => 'myform'))!!}
@@ -68,16 +69,20 @@
                                     <th>IVA</th>
                                     <th>Motivo</th>
                                     <th>Opciones</th>
+
                                 </tr>
                             </thead>
                             <tbody>
                             {{--Validacion de campos , aunque se valida desde el controlador, tambien tiene una validacion adicional
                             en la vista --}}
-                                <?php  use Carbon\Carbon; $registro = 1; $aux = 0;
+
+                                <?php
+                                use Carbon\Carbon;
+                                $registro = 1; $aux = 0;
                                 $tituloov = null; $titulotc = null; $titulomo = null; $titulobr = null;
                                 $titulode = null; $titulovd = null; $titulorz = null; $titulofe = null;
-                                $colorov = null; $colortc = null; $colormo = null; $colorbr = null;
-                                $colorde = null; $colorvd = null; $colorrz = null; $colorfe = null; ?>
+                                $checkbox = 'enable';
+                                ?>
 
                                 @foreach( $facturas as $fac)
 
@@ -86,105 +91,110 @@
                                 <?php $subtotal = $fac->bruto - $fac->descuento;  ?>
                                 <?php $errors = 0; ?>
 
+                                @if($fac->codigo_alterno == null)
+                                <?php $errors++;  $titulohtr = '* Cliente no ha sido creado en DMS *'; ?>
+                                @else
+                                  <?php $titulohtr = null; ?>
+                                @endif
 
-                                    @if($fac->codigo_alterno == null)
-                                    <?php $errors++; $colthr = 'text-danger'; $titulohtr = 'Debe tener un tipo de cliente' ?>
-                                    @else
-                                      <?php $colthr = null; $titulohtr = null; ?>
-
-                                    @endif
-
-
-                                  {{--evalua que tenga un tipo de cliente --}}
-                                  @if ($fac->tipo_cliente == null)
-                                    <?php $errors++; $colortc = 'text-danger'; $titulotc = 'Debe tener un tipo de cliente' ?>
-                                  @else
-                                    <?php $colortc = null; $titulotc = null; ?>
-                                  @endif
+                                {{--evalua que tenga un tipo de cliente --}}
+                                @if ($fac->tipo_cliente == null)
+                                  <?php $errors++; $titulotc = '* Debe tener un tipo de cliente *'; ?>
+                                @else
+                                  <?php $titulotc = null; ?>
+                                @endif
 
                                   {{--evalua que la Factura tenga un valor bruto mayor a 0 --}}
-                                  @if ($fac->bruto == 0)
-                                    <?php $errors++; $colorbr = 'text-danger'; $titulobr = 'El Valor no debe ser 0' ?>
-                                  @else
-                                    <?php $colorbr = null; $titulobr = null; ?>
-                                  @endif
+                                @if ($fac->bruto == 0)
+                                  <?php $errors++; $titulobr = '* El Valor no debe ser 0 *'; ?>
+                                @else
+                                  <?php $titulobr = null; ?>
+                                @endif
 
                                   {{--evalua que la Factura no tenga un valor bruto menor a 3000 --}}
-                                  @if ($fac->bruto < 3000)
-                                    <?php $errors++; $colorbr = 'text-danger'; $titulobr = 'El Valor es demasiado bajo' ?>
-                                  @else
-                                    <?php $colorbr = null; $titulobr = null; ?>
-                                  @endif
+                                @if ($fac->bruto < 3000)
+                                  <?php $errors++; $titulobr = '* El Valor es demasiado bajo *'; ?>
+                                @else
+                                  <?php $titulobr = null; ?>
+                                @endif
 
                                   {{--evalua que la Factura no tenga un valor bruto mayor a 20.000.000 --}}
-                                  @if ($fac->bruto > 20000000)
-                                    <?php $errors++; $colorbr = 'text-danger'; $titulobr = 'El Valor es demasiado alto' ?>
-                                  @else
-                                    <?php $colorbr = null; $titulobr = null; ?>
-                                  @endif
+                                @if ($fac->bruto > 20000000)
+                                  <?php $errors++;  $titulobr = '* El Valor es demasiado alto *'; ?>
+                                @else
+                                  <?php $titulobr = null; ?>
+                                @endif
 
                                   {{--evalua que la Factura no tenga un porcentaje de descuento mayor a 25% --}}
-                                  @if ($pordesc > 25)
-                                    <?php $errors++; $colorde = 'text-danger'; $titulode = 'El descuento es demasiado alto' ?>
-                                  @else
-                                    <?php $colorde = null; $titulode = null; ?>
-                                  @endif
+                                @if ($pordesc > 25)
+                                  <?php $errors++; $titulode = '* El descuento es demasiado alto *'; ?>
+                                @else
+                                  <?php $titulode = null; ?>
+                                @endif
 
                                   {{--evalua que la Factura tenga razon social --}}
-                                  @if ($fac->razon_social == null)
-                                    <?php $errors++; $colorrz = 'text-danger'; $titulorz = 'La Factura debe tener Razon Social' ?>
-                                  @else
-                                    <?php $colorfe = null; $titulofe = null; ?>
-                                  @endif
+                                @if ($fac->razon_social == null)
+                                  <?php $errors++; $titulorz = '* La Factura debe tener Razon Social *'; ?>
+                                @else
+                                  <?php $titulofe = null; ?>
+                                @endif
 
                                   {{--evalua que la Factura tenga Fecha --}}
-                                  @if ($fac->fecha == null)
-                                    <?php $errors++; $colorfe = 'text-danger'; $titulofe = 'La Factura debe tener fecha' ?>
-                                  @else
-                                    <?php $colorfe = null; $titulofe = null; ?>
-                                  @endif
+                                @if ($fac->fecha == null)
+                                  <?php $errors++; $titulofe = '* La Factura debe tener fecha *'; ?>
+                                @else
+                                  <?php $titulofe = null; ?>
+                                @endif
 
                                   {{--Muestra si es la primera factura --}}
-                                  @if ($registro == 1)
-                                    <?php $colornu = 'text-danger'; $titulonu = 'Este es el primer Registro'?>
-                                  @else
-                                    <?php $colornu = null;; $titulonu = null; ?>
-                                  @endif
+                                @if ($registro == 1)
+                                  <?php $titulonu = '* Este es el primer Registro *'; ?>
+                                @else
+                                  <?php $titulonu = null; ?>
+                                @endif
 
                                     {{--evalua que el consecutivo sea correcto --}}
-                                  @if ($fac->numero - $aux > 1)
-                                    <?php $errors++; $colornu = 'text-danger'; $titulonu = 'Error en el consecutivo'   ?>
-                                  @else
-                                    <?php $colornu = ''; $titulonu = '' ?>
-                                  @endif
+                                @if ($fac->numero - $aux > 1)
+                                  <?php $errors++; $titulonu = '* Error en el consecutivo *'; ?>
+                                @else
+                                  <?php $titulonu = null; ?>
+                                @endif
 
+                                @if ($errors >= 1)
+                                  <?php $checkbox = 'disabled'; ?>
+                                  <?php $colthr = 'text-danger'; ?>
+                                @else
+                                  <?php $checkbox = 'enable'; ?>
+                                  <?php $colthr = null; ?>
+                                @endif
+
+                                <?php $errortex = ' '.$titulohtr.' '.$titulotc.' '.$titulobr.' '.$titulonu.' '.$titulode.' '.$titulorz.' '.$titulofe ?>
 
                                   <tr class="<?php echo $colthr ?>">
-                                      <td><input type="checkbox" class="checkboxes" id="CHK{{ $fac->numero }}" name="CHK{{$fac->numero}}"/></td >
-                                      <td class=" <?php echo $colornu ?>" title="<?php echo $titulonu ?>" >{{ $fac->numero}}</td>
-                                      <td class=" <?php echo $colorov ?>" title="<?php echo $tituloov ?>"> {{ $fac->ov }}</td>
-                                      <td class=" <?php echo $colorfe ?>" title="<?php echo $titulofe ?>"> {{ Carbon::parse( $fac->fecha)->format('Y-m-d') }}</td>
-                                      <td>{{$fac->plazo}}</td>
-                                      <td class=" <?php echo $colorrz ?>" title="<?php echo $titulorz ?>"> {{ $fac->razon_social}} </td>
-                                      <td class=" <?php echo $colortc ?>" title="<?php echo $titulotc ?>"> {{ $fac->tipo_cliente }} </td>
-                                      <td class=" <?php echo $colorvd ?>" title="<?php echo $titulovd ?>"> {{ $fac->nomvendedor}} </td>
-                                      <td class=" <?php echo $colorbr ?>" title="<?php echo $titulobr ?>"> {{ number_format($fac->bruto,0,',','.') }}</td>
-                                      <td>{{ number_format($fac->descuento,0,',','.' )}}</td>
-                                      <td class=" <?php echo $colorde ?>" title="<?php echo $titulode ?>"> {{ number_format($pordesc ,'0',',','.') }}</td>
-                                      <td>{{ number_format($fac->valor_iva,0,',','.' )}}</td>
-                                      <td class=" <?php echo $colormo ?>" title="<?php echo $titulomo ?>"> {{ $fac->motivo }}</td>
+                                      <td><input type="checkbox" class="checkboxes" id="CHK{{ $fac->numero }}" name="CHK{{$fac->numero}}" <?php echo $checkbox ?> ></td >
+                                      <td> {{ $fac->numero}}</td>
+                                      <td> {{ $fac->ov }}</td>
+                                      <td> {{ Carbon::parse( $fac->fecha)->format('Y-m-d') }}</td>
+                                      <td> {{ $fac->plazo}}</td>
+                                      <td> {{ $fac->razon_social}} </td>
+                                      <td> {{ $fac->tipo_cliente }} </td>
+                                      <td> {{ $fac->nomvendedor}} </td>
+                                      <td> {{ number_format($fac->bruto,0,',','.') }}</td>
+                                      <td> {{ number_format($fac->descuento,0,',','.' )}}</td>
+                                      <td> {{ number_format($pordesc ,'0',',','.') }}</td>
+                                      <td> {{ number_format($fac->valor_iva,0,',','.' )}}</td>
+                                      <td> {{ $fac->motivo }}</td>
                                       <td>
                                           <div class="btn-group ml-auto float-right">
                                               @can('fact.edit')
-                                                  <button class="btn btn-sm btn-light" title="<?php echo $titulohtr ?>" disabled>
+                                                  <button class="btn btn-sm btn-light " id="BTN{{$fac->numero}}" data-toggle="tooltip" data-placement="top" title="<?php echo $errortex ?>" >
                                                     <?php echo $errors  ?>
                                                   </button>
+
                                                   <a href="{{ route('fe.edit', $fac->numero) }}" class="btn btn-sm btn-outline-light" id="edit-fac" >
                                                       <i class="far fa-edit"></i>
                                                   </a>
-                                                  <a href="#" class="btn btn-sm btn-outline-light" id="ver-fac">
-                                                      <i class="fa fa-eye"></i>
-                                                  </a>
+
                                               @endcan
                                                   <input type="hidden" name="numero{{ $fac->numero }}" value="{{ $fac->numero }}" id="numero{{ $fac->numero }}"  >
                                           </div>
@@ -200,204 +210,26 @@
             </div>
         </div>
     </div>
-    <iframe id="my_iframe" style="display:none;"></iframe>
+
 @stop
     {{Form::close()}}
 
-
 @push('javascript')
     <script>
-        $( function() {
-            var from = $( "#fechaInicial" )
-                .datepicker({
-                    dateFormat: "yymmdd 00:00:00",
-                    changeMonth: true,
-                    changeYear: false,
-                    closeText: 'Cerrar',
-                    prevText: 'Ant',
-                    nextText: 'Sig',
-                    currentText: 'Hoy',
-                    monthNames: ['Enero', 'Febrero', 'Marzo', 'Abril', 'Mayo', 'Junio', 'Julio', 'Agosto', 'Septiembre', 'Octubre', 'Noviembre', 'Diciembre'],
-                    monthNamesShort: ['Ene','Feb','Mar','Abr', 'May','Jun','Jul','Ago','Sep', 'Oct','Nov','Dic'],
-                    dayNames: ['Domingo', 'Lunes', 'Martes', 'Miércoles', 'Jueves', 'Viernes', 'Sábado'],
-                    dayNamesShort: ['Dom','Lun','Mar','Mié','Juv','Vie','Sáb'],
-                    dayNamesMin: ['Do','Lu','Ma','Mi','Ju','Vi','Sá'],
-                    weekHeader: 'Sm',
-                    firstDay: 1,
-                    isRTL: false,
-                    showMonthAfterYear: false,
-                    yearSuffix: '',
-                    showAnim: "drop"
-                })
-                .on( "change", function() {
-                    to.datepicker( "option", "minDate", getDate( this ) );
-                }),
-                to = $( "#fechaFinal" ).datepicker({
-                    dateFormat: "yymmdd 23:59:59",
-                    changeMonth: true,
-                    changeYear: false,
-                    closeText: 'Cerrar',
-                    prevText: 'Ant',
-                    nextText: 'Sig',
-                    currentText: 'Hoy',
-                    monthNames: ['Enero', 'Febrero', 'Marzo', 'Abril', 'Mayo', 'Junio', 'Julio', 'Agosto', 'Septiembre', 'Octubre', 'Noviembre', 'Diciembre'],
-                    monthNamesShort: ['Ene','Feb','Mar','Abr', 'May','Jun','Jul','Ago','Sep', 'Oct','Nov','Dic'],
-                    dayNames: ['Domingo', 'Lunes', 'Martes', 'Miércoles', 'Jueves', 'Viernes', 'Sábado'],
-                    dayNamesShort: ['Dom','Lun','Mar','Mié','Juv','Vie','Sáb'],
-                    dayNamesMin: ['Do','Lu','Ma','Mi','Ju','Vi','Sá'],
-                    weekHeader: 'Sm',
-                    firstDay: 1,
-                    isRTL: false,
-                    showMonthAfterYear: false,
-                    yearSuffix: '',
-                    showAnim: "drop"
-                })
-                    .on( "change", function() {
-                        from.datepicker( "option", "maxDate", getDate( this ) );
-                    });
-            function getDate( element ) {
-                var date;
-                var dateFormat = "yymmdd";
-                try {
-                    date = $.datepicker.parseDate( dateFormat, element.value );
-                } catch( error ) {
-                    date = null;
-                }
-                return date;
-            }
-        });
-        {{--- fin de datepicker --}}
-
-
-      {{--Inicio datatables--}}
-        $(document).ready(function () {
-            $("#tfac").DataTable({
-                order: [],
-                columns: [
-                    // permite ordenar columnas y le da el abributo buscar
-                    {"orderable": false, "searchable": false},
-                    {"orderable": true, "searchable": true},
-                    {"orderable": false, "searchable": true},
-                    {"orderable": false, "searchable": false},
-                    {"orderable": true, "searchable": true},
-                    {"orderable": true, "searchable": true},
-                    {"orderable": true, "searchable": false},
-                    {"orderable": true, "searchable": true},
-                    {"orderable": false, "searchable": false},
-                    {"orderable": false, "searchable": false},
-                    {"orderable": false, "searchable": false},
-                    {"orderable": false, "searchable": false},
-                    {"orderable": true, "searchable": true},
-                    {"orderable": false, "searchable": false},
-                ],
-
-                columnDefs: [
-                    {
-                        targets: 0,
-                    }
-                ],
-
-                select: {
-                    style: 'multi'
-                },
-
-                language: {
-                    // traduccion de datatables
-                    processing: "Procesando...",
-                    search: "Buscar&nbsp;:",
-                    lengthMenu: "Mostrar _MENU_ registros",
-                    info: "Mostrando registros del _START_ al _END_ de un total de _TOTAL_ registros",
-                    infoEmpty: "Mostrando registros del 0 al 0 de un total de 0 registros",
-                    infoFiltered: "(filtrado de un total de _MAX_ registros)",
-                    infoPostFix: "",
-                    loadingRecords: "Cargando...",
-                    zeroRecords: "No se encontraron resultados",
-                    emptyTable: "Ningún registro disponible en esta tabla :C",
-                    paginate: {
-                        first: "Primero",
-                        previous: "Anterior",
-                        next: "Siguiente",
-                        last: "Ultimo"
-                    },
-                    aria: {
-                        sortAscending: ": Activar para ordenar la columna de manera ascendente",
-                        sortDescending: ": Activar para ordenar la columna de manera descendente"
-                    }
-                }
-            });
-        });
-
-        $(document).ready(function () {
-            $("#CrearXml").click(function () {
-                var selected = [];
-                $(".checkboxes").each(function () {
-                    if (this.checked) {
-                        var numero = this.id.substring(3,9);
-                        var factura ={
-                            "numero":numero,
-                        };
-                        selected.push(factura);
-                    }
-                });
-                if (selected.length) {
-                    $.ajaxSetup({
-                        headers: {
-                            'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
-                        }
-                    });
-                    toastr.info('Un momento por favor...','Generando XML.',{
-                        "closeButton": false,
-                        "debug": false,
-                        "newestOnTop": false,
-                        "progressBar": false,
-                        "preventDuplicates": false,
-                        "onclick": null,
-                        "showDuration": "300",
-                        "hideDuration": "10000",
-                        "timeOut": "5000",
-                        "extendedTimeOut": "1000",
-                        "showEasing": "swing",
-                        "hideEasing": "linear",
-                        "showMethod": "fadeIn",
-                        "hideMethod": "fadeOut"
-                    });
-
-                    $.ajax({
-                        cache: false,
-                        type: 'post',
-                        dataType: 'json', // importante para que
-                        data: {selected: JSON.stringify(selected)}, // jQuery convierta el array a JSON
-                        url: 'fe/xml',
-                         success: function () {
-                           toastr.success("El Archivo XML se ha generado con Exito!.");
-                          // preventDefault();  //stop the browser from following
-
-                           window.open('XML/Facturacion_electronica_Facturas.xml');
-
-                         }
-                    });
-                } else
-                  toastr.error("Debes seleccionar al menos una Factura.");
-                return false;
-            });
-
-        });
-
-        $("#selectAll").click(function(){
-            $("input[type=checkbox]").prop('checked', $(this).prop('checked'));
-        });
+			$(document).ready(function() {
+				$('[data-toggle="tooltip"]').tooltip();
+			});
     </script>
 
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
+    <script type="text/javascript" src="/JsGlobal/FE/index.js" ></script>
     <link rel="stylesheet" href="//code.jquery.com/ui/1.12.1/themes/base/jquery-ui.css">
     <script src="https://code.jquery.com/jquery-1.12.4.js"></script>
     <script src="https://code.jquery.com/ui/1.12.1/jquery-ui.js"></script>
     <link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/v/bs4/dt-1.10.18/datatables.min.css"/>
     <script type="text/javascript" src="https://cdn.datatables.net/v/bs4/dt-1.10.18/datatables.min.js"></script>
     <link type="text/css" href="//gyrocode.github.io/jquery-datatables-checkboxes/1.2.11/css/dataTables.checkboxes.css" rel="stylesheet" />
-    <script type="text/javascript" src="//gyrocode.github.io/jquery-datatables-checkboxes/1.2.11/js/dataTables.checkboxes.min.js"></script>
-    <link type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/awesome-bootstrap-checkbox/0.3.7/awesome-bootstrap-checkbox.css" rel="stylesheet">
-    <link rel="stylesheet" type="text/css" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
-
-
 @endpush
